Add default value for level in Result
Once you have a log-frame object you still can't actually use the
UI to add Result objects because the first one at the top of the
tree doesn't specify its "level" as zero. By making this the default
value in the model we can get round this and start filling in the
logframe as soon as it is created. Before that we had to create a
top level result manuall.